requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
ecdsa==0.19.1
fastapi==0.124.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
bcrypt==4.2.1
python-dotenv==1.2.1
python-jose==3.3.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
